Research_paper_explorer_RAG/requirements.txt
chromadb
sentence-transformers
pypdf
openai
python-dotenv
numpy<2.0
streamlit
requests